[chore][pkg/stanza] Cleanup recombine operator files
