+++ pkg/stanza/operator/transformer/recombine/transformer_test.go
